+++ vvm_to_signalk/config_decoder.py
